use trove devstack plugin  from branch being built
The trove plugin was always installed from master branch.
This change allows us to use the plugin as defined in the trove
branch we are building.

The plugins installation installs to /opt/stack/new just like the
other pieces of code.  It seems safer to checkout the same branch
for plugins as we do for the actual Trove code in test.

Change-Id: I3f9c2e72a99785db641dc4198d99927dff80197c